requirements.txt
delta-spark==1.0.0
pylint==2.14.1
pandas==1.2.0
